SIH-2019/requirements.txt
boto3==1.9.96
botocore==1.12.96
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-filter==2.1.0
django-multiselectfield==0.1.8
django-otp==0.5.2
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.14
idna==2.8
jmespath==0.9.3
Markdown==3.0.1
Pillow==5.4.1
PyJWT==1.7.1
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
twilio==6.24.1
urllib3==1.24.1
